server/requirements.txt
# ============================================================
#  MIS — Python Dependencies
# ============================================================

# ── Data Processing ──────────────────────────────────────────
pandas==2.3.3
openpyxl==3.1.5
xlrd==2.0.1
pdfplumber==0.11.4

# ── Core Framework ───────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6

# ── Settings / Environment ───────────────────────────────────
pydantic==2.9.2
pydantic-settings==2.5.2

# ── Database ─────────────────────────────────────────────────
sqlmodel==0.0.21
sqlalchemy==2.0.35
aiosqlite==0.20.0

# ── Authentication ───────────────────────────────────────────
bcrypt==4.2.0
python-multipart==0.0.12
itsdangerous==2.2.0

# ── Admin UI ─────────────────────────────────────────────────
sqladmin==0.19.0



# ── Templating ───────────────────────────────────────────────
jinja2==3.1.4

# ── HTTP / Serialization ─────────────────────────────────────
httpx==0.27.2
orjson==3.10.7

# ── Additional Utilities ─────────────────────────────────────
psycopg2-binary>=2.9.9
python-dotenv
